+++ tap/api/go.mod
@@ -3,7 +3,6 @@ module github.com/up9inc/mizu/tap/api
 go 1.17
 
 require (
-	github.com/google/gopacket v1.1.19
 	github.com/google/martian v2.1.0+incompatible
 	github.com/up9inc/mizu/shared v0.0.0
 )
